Move verify command to its own files
